Created a place to test bits of the graphical interface in isolation.
