+++ 2_adventures/001/jbrains/11/try_connecting_grid_square_to_model.rb
@@ -1,0 +1,12 @@
+require "./langtons_ant_ui"
+
+Gtk.init
+
+main_window = Gtk::Window.new
+main_window.set_title("Flip the color of a grid square")
+main_window.signal_connect("destroy") {
+  Gtk.main_quit
+}
+main_window.show_all
+
+Gtk.main
